rename storage interface to base_storage and move to common/storage
