+++ tests/iam/inventory_tests/crawling_test.py
@@ -18,7 +18,7 @@ import unittest
 
 
 from tests.unittest_utils import ForsetiTestCase
-from google.cloud.security.iam.inventory.inventory2.storage import Memory as MemoryStorage
+from google.cloud.security.common.storage.base_storage import Memory as MemoryStorage
 from google.cloud.security.iam.inventory.inventory2.progress import Progresser
 from google.cloud.security.iam.server import InventoryConfig
 from google.cloud.security.inventory.crawler import run_crawler
